requirements.txt
# Neural Retriever-Reranker RAG Pipeline Dependencies

# Core ML/Data Science
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Deep Learning and Transformers
torch>=1.9.0
sentence-transformers>=2.2.0
transformers>=4.20.0

# Amazon STaRK SKB
huggingface_hub 
tqdm 
click

# FAISS for dense retrieval
faiss-cpu>=1.7.0  # Use faiss-gpu if GPU support needed

# BM25 for sparse retrieval
rank-bm25>=0.2.1

# Lightning IR for Webis rerankers (FRWSR only)
lightning-ir>=0.1.0

# Text processing
nltk>=3.7
regex>=2022.1.18

# Evaluation and metrics
scipy>=1.7.0
tqdm>=4.62.0

# Configuration and utilities
pyyaml>=6.0
click>=8.0.0

# Development and testing
pytest>=6.2.0
pytest-cov>=3.0.0
black>=22.0.0
flake8>=4.0.0
mypy>=0.910

# Visualization (optional)
matplotlib>=3.5.0
seaborn>=0.11.0

# Jupyter notebooks (optional)
jupyter>=1.0.0
ipykernel>=6.0.0

# Data handling
pyarrow>=5.0.0  # For faster CSV operations

# Specific versions matching actual implementations
# sentence-transformers for E5-Large embeddings (FRMR/FRWSR)
# cross-encoder models for MS MARCO reranking (FRMR/BARMR)
# rank-bm25 for sparse retrieval (BARMR)
# lightning-ir for Webis Set-encoder (FRWSR)